Updated Artefact Hud Aspect
I have added more to the artefact aspect of the hud. We can still use the original if people would like to.

I mainly just stylised it to the theme of the hud and made it easier for people to edit in the future, and program within their game levels.

If you wish to program said artefact hud part just cast to the "ThirdPersonCharacter" blueprint. Get the instances of the "as-hud" then call the function "get-artefact-rtxt-values".